services/ai_examiner/requirements.txt
# 2025-10-22: Dependencies for AI Examiner Microservice
# Authors: Cascade

Flask==3.0.0
openai==1.3.3
python-dotenv==1.0.0
